+++ packages/sdk/src/util/api-factory.ts
@@ -3,16 +3,11 @@ import * as Keystore from '@meeco/keystore-api-sdk';
 import * as Vault from '@meeco/vault-api-sdk';
 import { Configuration } from '@meeco/vault-api-sdk';
 import * as VC from '@meeco/vc-api-sdk';
+import chalk from 'chalk';
 import { debug } from 'debug';
 import { Environment } from '../models/environment';
 import { IIdentityNetworkToken, IKeystoreToken, IVaultToken, IVCToken } from '../services/service';
 import SDKFormData from './sdk-form-data';
-
-/**
- * INFO: using 'import' statement causes typescript errors either in tests or in built version of the package
- */
-/* tslint:disable no-var-requires */
-const chalk = require('chalk');
 
 let fetchLib = fetch;