+++ packages/sdk/src/util/sdk-form-data.ts
@@ -1,5 +1,4 @@
-// tslint:disable-next-line: no-var-requires
-const FormData = require('form-data');
+import FormData from 'form-data';
 
 /**
  * @hidden
@@ -10,12 +9,12 @@ const FormData = require('form-data');
  * This monkey-patches the constructor of FormData to ensure the argument always exists
  */
 class SDKFormData extends FormData {
-  append(...args) {
-    if (args[0] === 'attachment[file]' || args[0] === 'binary[file]') {
-      return super.append(args[0], args[1], 'file');
+  append(key: string, value: any, options?: FormData.AppendOptions | string) {
+    if (key === 'attachment[file]' || key === 'binary[file]') {
+      return super.append(key, value, 'file');
     }
 
-    return super.append(...args);
+    return super.append(key, value, options);
   }
 }